career_agent_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
dashscope==1.14.0
neo4j==5.14.1
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.0.3
openpyxl==3.1.2
tqdm==4.66.1
cpca==0.5.5
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0